chore: codebase quality sweep — bug fixes, dead-code removal, file org
Autonomous /improve-codebase sweep focused on config/clients/warnings/
tools/publish-site/upgrade-migrations.

Bug fixes:
- extractchecksum: TrimLeft->TrimPrefix so dotfile basenames keep their
  leading dot ("./" is a path marker; TrimLeft stripped it as a cutset).
- publish-site: de-duplicate colliding H2 slugs to match Docusaurus
  github-slugger anchors, including the suffix-collision edge case
  (foo, foo-1, foo-1-1).

Dead-code removal (zero callers):
- internal/config: delete commands.go, instructions.go and their tests;
  remove unused LoadConfig loader. Move utf8BOM/osReadFileFunc to skills.go.

File organization (no behavior change):
- internal/install: split skills-migration code out of
  upgrade_migrations.go into upgrade_migrations_skills.go.

Fail-loud / determinism / test quality:
- skills.go: reject duplicate front-matter metadata keys (parity with
  top-level keys; avoids silent last-value-wins).
- clients claude/codex: best-effort stderr warning write.
- messages: centralize config string constants.
- determinism/exactness improvements across version/warnings/probe/
  launchers/dispatch/install tests.

Memory:
- ISSUES.md: log deferred findings (secret-in-URL precision/recall
  decision reverted pending a human call; fs-os-abstraction and semver
  consolidation architectural items).
- BACKLOG.md: unrelated pre-existing user entry rides along.
