+++ internal/clients/codex/launch_test.go
@@ -48,10 +48,7 @@ func TestLaunchCodex(t *testing.T) {
 	cfg := &config.ProjectConfig{
 		Config: config.Config{
 			Agents: config.AgentsConfig{
-				Codex: config.CodexConfig{
-					Model:           "model",
-					ReasoningEffort: "low",
-				},
+				Codex: config.CodexConfig{},
 			},
 		},
 		Root: root,
@@ -74,10 +71,7 @@ func TestLaunchCodexError(t *testing.T) {
 	cfg := &config.ProjectConfig{
 		Config: config.Config{
 			Agents: config.AgentsConfig{
-				Codex: config.CodexConfig{
-					Model:           "model",
-					ReasoningEffort: "low",
-				},
+				Codex: config.CodexConfig{},
 			},
 		},
 		Root: root,
@@ -124,28 +118,5 @@ func TestEnsureCodexHomeWarnsOnMismatch(t *testing.T) {
 	wantWarning := fmt.Sprintf(messages.ClientsCodexHomeWarningFmt, current, expected)
 	if !strings.Contains(stderr, wantWarning) {
 		t.Fatalf("expected stderr to contain warning %q, got %q", wantWarning, stderr)
-	}
-}
-
-func TestEnsureCodexHome_WarningWriteFailureLeavesEnvUnchanged(t *testing.T) {
-	root := t.TempDir()
-	current := filepath.Join(t.TempDir(), "other")
-	env := []string{"CODEX_HOME=" + current}
-
-	origStderr := os.Stderr
-	_, stderrWriter, err := os.Pipe()
-	if err != nil {
-		t.Fatalf("pipe: %v", err)
-	}
-	if err := stderrWriter.Close(); err != nil {
-		t.Fatalf("close pipe writer: %v", err)
-	}
-	os.Stderr = stderrWriter
-	t.Cleanup(func() { os.Stderr = origStderr })
-
-	out := ensureCodexHome(root, env)
-	value, ok := clients.GetEnv(out, "CODEX_HOME")
-	if !ok || value != current {
-		t.Fatalf("expected CODEX_HOME to remain unchanged, got %q", value)
 	}
 }
