--- internal/config/commands.go
@@ -1,33 +1,0 @@
-package config
-
-import (
-	"bufio"
-	"fmt"
-	"os"
-	"strings"
-
-	"github.com/conn-castle/agent-layer/internal/messages"
-)
-
-// LoadCommandsAllow reads .agent-layer/commands.allow into a slice of prefixes.
-func LoadCommandsAllow(path string) ([]string, error) {
-	data, err := os.ReadFile(path) // #nosec G304 -- path is the caller-resolved .agent-layer/commands.allow location (derived via paths.AgentLayerDir()), not user input.
-	if err != nil {
-		return nil, fmt.Errorf(messages.ConfigMissingCommandsAllowlistFmt, path, err)
-	}
-
-	var commands []string
-	scanner := bufio.NewScanner(strings.NewReader(string(data)))
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-		if line == "" || strings.HasPrefix(line, "#") {
-			continue
-		}
-		commands = append(commands, line)
-	}
-	if err := scanner.Err(); err != nil {
-		return nil, fmt.Errorf(messages.ConfigFailedReadCommandsAllowlistFmt, path, err)
-	}
-
-	return commands, nil
-}
